+++ src/main/java/io/ix0rai/rainglow/config/RainglowConfigScreen.java
@@ -6,10 +6,14 @@ import dev.lambdaurora.spruceui.option.SpruceOption;
 import dev.lambdaurora.spruceui.option.SpruceSimpleActionOption;
 import dev.lambdaurora.spruceui.screen.SpruceScreen;
 import dev.lambdaurora.spruceui.widget.SpruceButtonWidget;
-import dev.lambdaurora.spruceui.widget.container.SpruceOptionListWidget;
+import dev.lambdaurora.spruceui.widget.SpruceLabelWidget;
 import io.ix0rai.rainglow.Rainglow;
+import io.ix0rai.rainglow.SquidColour;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.text.Style;
+import net.minecraft.text.Text;
+import net.minecraft.util.Language;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
@@ -62,10 +66,17 @@ public class RainglowConfigScreen extends SpruceScreen {
         super.init();
 
         int buttonHeight = 20;
+        int buttonOffset = 30;
 
-        SpruceOptionListWidget options = new SpruceOptionListWidget(Position.of(0, 22), this.width, this.height - (35 + 22));
-        options.addOptionEntry(this.modeOption, this.customOption);
-        this.addDrawableChild(options);
+        this.addDrawableChild(this.modeOption.createWidget(Position.of(this.width / 2 - 205, this.height / 6 - buttonHeight + buttonOffset), 200));
+        this.addDrawableChild(this.customOption.createWidget(Position.of(this.width / 2 + 5, this.height / 6 - buttonHeight + buttonOffset), 200));
+
+        this.addDrawableChild(new SpruceLabelWidget(Position.of(this, 0, this.height / 9), Rainglow.translatableText("config.title"), this.width, true));
+        StringBuilder text = new StringBuilder(Language.getInstance().get(Rainglow.translatableTextKey("config.current_colours")));
+        for (SquidColour colour : Rainglow.CONFIG.getMode().getColours()) {
+            text.append("\n").append(Language.getInstance().get(Rainglow.translatableTextKey("colour." + colour.getId())));
+        }
+        this.addDrawableChild(new SpruceLabelWidget(Position.of(this, 0, this.height / 4 + buttonHeight), Text.of(text.toString()).copy().setStyle(Style.EMPTY.withColor(Rainglow.CONFIG.getMode().getText().getStyle().getColor())), this.width, true));
 
         this.addDrawableChild(this.resetOption.createWidget(Position.of(this, this.width / 2 - 155, this.height - 29), 150));
         this.addDrawableChild(new SpruceButtonWidget(Position.of(this, this.width / 2 - 155 + 160, this.height - 29), 150,